test_project/requirement.txt
Django
Pillow
psycopg2-binary
python-dotenv
algoliasearch-Django
djangorestframework
pyyaml
requests
django-cors-headers
